django_capstone/requirements.txt
social-auth-app-django
psycopg2
requests
python-dateutil
twitch-python
pytz
django-background-tasks